refactor: drop embedded FrozenList
The FrozenList class here was long ago split out into its own tiny
shared library. Drop the embedded implementation, and the associated
tests, in favor of that library. Related conversion helpers were moved
to the 'convert' module.
