+++ pubtools/pulplib/_impl/model/unit/modulemd.py
@@ -4,10 +4,7 @@ from .base import Unit, unit_type
 
 from ..attr import pulp_attrib
 from ... import compat_attr as attr
-from ..frozenlist import (
-    frozenlist_or_none_converter,
-    frozenlist_or_none_sorted_converter,
-)
+from ..convert import frozenlist_or_none_converter, frozenlist_or_none_sorted_converter
 
 
 @unit_type("modulemd")
